♻️refactor: SignUp step 추가
- 생년월일 필수 입력을 위해 생년월일 스탭을 별도로 분리
