+++ src/app/signup/sections/SignUpStep3.tsx
@@ -1,38 +1,89 @@
-import React from "react";
-import { UseFormGetValues, UseFormRegister } from "react-hook-form";
+"use client";
 
-import { UserIntensityItem } from "@/components";
+import React, { useEffect } from "react";
+import { FieldErrors, UseFormSetValue, UseFormRegister, UseFormGetValues } from "react-hook-form";
 
+import useTheme from "@/lib/hooks/useTheme";
+
+import { Button, UserEditInput } from "@/components";
 import { SignUpFormProps } from "../SignUp.controller";
+import useSignUpModel from "../SignUp.model";
 import * as S from "../SignUp.styles";
-import { INTENSITY_OPTIONS } from "@/lib/constants/variable";
+
+import { SEX_OPTIONS } from "@/lib/constants/variable";
+import { USER_INPUT_PLACEHOLDER, validation_user } from "@/lib/constants/userConstants";
 
 interface SignUpStep3Props {
-  register: UseFormRegister<SignUpFormProps>;
   getValues: UseFormGetValues<SignUpFormProps>;
+  setValue: UseFormSetValue<SignUpFormProps>;
+  register: UseFormRegister<SignUpFormProps>;
+  errors: FieldErrors<SignUpFormProps>;
 }
 
-const SignUpStep3 = ({ getValues, register }: SignUpStep3Props) => {
-  const selectedIntensity = getValues("exerciseIntensity");
+const SignUpStep3 = ({ getValues, setValue, register, errors }: SignUpStep3Props) => {
+  const theme = useTheme();
+
+  useEffect(() => {
+    const initState = getValues("sex");
+    if (initState) {
+      setSelectedSex(initState);
+    }
+  }, []);
+
+  const { selectedSex, setSelectedSex } = useSignUpModel();
+  useEffect(() => {
+    if (selectedSex) setValue("sex", selectedSex);
+  }, [setValue, selectedSex]);
+
+  const handleSexSelect = (sex: "MALE" | "FEMALE") => {
+    setSelectedSex(sex);
+    setValue("sex", sex);
+  };
 
   return (
-    <div className="w-full h-full">
-      <S.ExerciseIntensityTitleSection>
-        <S.Title style={{ marginBottom: "1.2rem" }}>평소 운동 강도</S.Title>
-        기초 대사량 계산에 필요한 활동 계수를 설정해요.
-      </S.ExerciseIntensityTitleSection>
+    <div className="h-full">
+      <S.GenderSection>
+        <S.Title>성별</S.Title>
+        <S.GenderButtonGroup>
+          {SEX_OPTIONS.map((option) => (
+            <Button
+              key={option.value}
+              onClick={() => handleSexSelect(option.value)}
+              style={{
+                width: "50%",
+                backgroundColor:
+                  selectedSex === option.value ? theme?.green_500 : theme?.transparent_10,
+                cursor: "pointer",
+              }}
+            >
+              {option.label}
+            </Button>
+          ))}
+        </S.GenderButtonGroup>
+      </S.GenderSection>
 
-      <S.ExerciseIntensityOptionList>
-        {INTENSITY_OPTIONS.map(({ label, value }) => (
-          <UserIntensityItem
-            key={value}
-            value={value}
-            optionDescription={label}
-            isSelected={selectedIntensity === value}
-            register={register("exerciseIntensity")}
+      <S.PhysicalSection>
+        <S.PhysicalGroup>
+          <UserEditInput
+            title={"키"}
+            unit={"cm"}
+            inputType={"number"}
+            placeholder={USER_INPUT_PLACEHOLDER.HEIGHT}
+            errorsMessage={errors.height && errors.height.message}
+            register={register("height", validation_user.height)}
           />
-        ))}
-      </S.ExerciseIntensityOptionList>
+        </S.PhysicalGroup>
+        <S.PhysicalGroup>
+          <UserEditInput
+            title={"몸무게"}
+            unit={"kg"}
+            inputType={"number"}
+            placeholder={USER_INPUT_PLACEHOLDER.WEIGHT}
+            errorsMessage={errors.weight && errors.weight.message}
+            register={register("weight", validation_user.weight)}
+          />
+        </S.PhysicalGroup>
+      </S.PhysicalSection>
     </div>
   );
 };
