+++ src/app/signup/sections/index.ts
@@ -2,3 +2,4 @@ export { default as SignUpStep1 } from "./SignUpStep1";
 export { default as SignUpStep2 } from "./SignUpStep2";
 export { default as SignUpStep3 } from "./SignUpStep3";
 export { default as SignUpStep4 } from "./SignUpStep4";
+export { default as SignUpStep5 } from "./SignUpStep5";
